src/requirements.txt
aioredis==1.3.1
appnope==0.1.2
asgi-redis==1.4.3
asgiref==3.3.4
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.4
backcall==0.2.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cycler==0.10.0
daphne==3.0.2
DateTime==4.3
decorator==4.4.2
defusedxml==0.7.1
dj-rest-auth==2.1.4
Django==3.1.7
django-allauth==0.44.0
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
dlib==19.21.1
drf-yasg==1.20.0
getch==1.0
gevent==20.12.1
greenlet==1.0.0
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
imagezmq==1.1.1
imutils==0.5.3
incremental==21.3.0
inflection==0.5.1
ipython==7.19.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.0
joblib==1.0.0
kiwisolver==1.3.1
MarkupSafe==2.0.0
matplotlib==3.3.3
moviepy==1.0.3
msgpack==1.0.2
msgpack-python==0.5.6
mysql-connector-python==8.0.23
networkx==2.5
numpy==1.19.5
oauthlib==3.1.0
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
packaging==20.9
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
proglog==0.1.9
prompt-toolkit==3.0.10
protobuf==3.14.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.4
PyJWT==2.0.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
PyWavelets==1.1.1
pyzmq==20.0.0
redis==2.10.6
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
scikit-image==0.18.1
scikit-learn==0.24.0
scipy==1.6.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
threadpoolctl==2.1.0
tifffile==2021.1.11
toml==0.10.2
tqdm==4.56.0
traitlets==5.0.5
Twisted==21.2.0
txaio==21.2.1
uritemplate==3.0.1
urllib3==1.26.2
wcwidth==0.2.5
websocket==0.2.1
websocket-client==0.57.0
zope.event==4.5.0
zope.interface==5.2.0
